+++ R/Shazam.R
@@ -127,6 +127,7 @@
 #'                          vertex_attr set_vertex_attr
 #' @importFrom  kedd        h.ucv
 #' @importFrom  KernSmooth  bkde
+#' @importFrom  ks          hucv kde
 #' @importFrom  lazyeval    interp
 #' @importFrom  MASS        fitdistr
 #' @importFrom  progress    progress_bar